+++ projects/play-with-migrations/src/server/config.ts
@@ -5,15 +5,16 @@ import { KnexDataProvider } from '../../../core/remult-knex/index.js'
 import Knex from 'knex'
 config()
 
-export const dataProvider = new KnexDataProvider(
-  Knex({
-    client: 'pg',
-    connection: process.env.DATABASE_URL,
-    //debug:true
-  }),
-)
+export const dataProvider =
+  // new KnexDataProvider(
+  //   Knex({
+  //     client: 'pg',
+  //     connection: process.env.DATABASE_URL,
+  //     //debug:true
+  //   }),
+  // )
 
-createPostgresDataProvider({
-  connectionString: process.env.DATABASE_URL,
-})
+  createPostgresDataProvider({
+    connectionString: process.env.DATABASE_URL,
+  })
 export const entities = [Task]