+++ projects/play-with-migrations/src/shared/tasks.ts
@@ -10,4 +10,6 @@ export class Task {
   title = ''
   @Fields.boolean()
   completed = false
+  @Fields.createdAt()
+  createdAt = new Date()
 }